refactor(ptx): update folder structure with splitting services
