+++ libs/ptx/src/index.ts
@@ -1,3 +1,6 @@
-export * from './api';
+export * from './activities';
 export * from './category';
+export * from './hotels';
+export * from './restaurants';
+export * from './scenes';
 export type { PTX } from './types';
